Extract converter implementations into a separate class.
